clean up user controller
