+++ web/controllers/user_controller.ex
@@ -1,19 +1,14 @@
 defmodule What3things.UserController do
   use What3things.Web, :controller
   alias What3things.{User, Email, Service, Mailer}
-  import Ecto.Query
 
   def create(conn, %{"user" => user_params}) do
     %{"name" => name,
       "postcode" => postcode,
       "age_range" => age_range} = user_params
 
-    existing_user = Repo.all(
-      from u in What3things.User,
-      where: u.name == ^name and u.postcode == ^postcode and u.age_range == ^age_range,
-      select: u
-    )
-
+    user_query = User.get_existing({ name, postcode, age_range })
+    existing_user = Repo.all(user_query)
     changeset = User.changeset(%User{}, user_params)
 
     case existing_user do
@@ -57,19 +52,8 @@ defmodule What3things.UserController do
 
   defp get_top3things(top3_ids) do
     top3_ids
-    |> format_ids
     |> Service.services_by_id
     |> Repo.all
-  end
-
-  defp format_ids(ids) do
-    if is_list(ids) do ids else ids |> ids_to_list  end
-  end
-
-  defp ids_to_list(ids) do
-    ids
-    |> String.split(",")
-    |> Enum.map(&String.to_integer/1)
   end
 
   def handle_email(%{email: email, service_ids: service_ids, uuid: uuid}) do
